Show share link in collapsed task header as well
